Break kanidm_tui into app,ui,messages modules
